Move convertor functions into their new boxes
