+++ libs/sdk-backend-bear/src/backend/workspace/elements/elements.ts
@@ -9,8 +9,7 @@ import { IAttributeElement, ObjRef } from "@gooddata/sdk-model";
 import invariant from "ts-invariant";
 
 import { BearAuthenticatedCallGuard } from "../../../types/auth";
-import { objRefToUri } from "../../../convertors/fromObjRef/api";
-import { getObjectIdFromUri } from "../../../utils/api";
+import { objRefToUri, getObjectIdFromUri } from "../../../utils/api";
 
 export class BearWorkspaceElements implements IElementQueryFactory {
     constructor(private readonly authCall: BearAuthenticatedCallGuard, public readonly workspace: string) {}
