+++ libs/sdk-backend-bear/src/backend/workspace/execution/executionResult.ts
@@ -18,8 +18,8 @@ import {
 import { IExecutionDefinition } from "@gooddata/sdk-model";
 import SparkMD5 from "spark-md5";
 import { BearAuthenticatedCallGuard } from "../../../types/auth";
-import { convertExecutionApiError } from "../../../errors/errorHandling";
-import { toAfmExecution } from "../../../convertors/toAfm/toAfmResultSpec";
+import { convertExecutionApiError } from "../../../utils/errorHandling";
+import { toAfmExecution } from "../../../convertors/toBackend/afm/ExecutionConverter";
 
 export class BearExecutionResult implements IExecutionResult {
     public readonly dimensions: IDimensionDescriptor[];
