+++ libs/sdk-backend-bear/src/backend/workspace/permissions/permissions.ts
@@ -2,7 +2,7 @@
 import { IWorkspacePermissionsFactory, IWorkspaceUserPermissions } from "@gooddata/sdk-backend-spi";
 import { IWorkspacePermissions, WorkspacePermission } from "@gooddata/sdk-model";
 import { BearAuthenticatedCallGuard } from "../../../types/auth";
-import { convertPermissions } from "../../../convertors/fromSdkModel/WorkspaceConverter";
+import { convertPermissions } from "../../../convertors/toBackend/WorkspaceConverter";
 
 const emptyPermissions = { permissions: {} };
