+++ libs/sdk-backend-bear/src/backend/workspace/users/index.ts
@@ -2,7 +2,7 @@
 import { IWorkspaceUsersQuery, IWorkspaceUsersQueryOptions, IWorkspaceUser } from "@gooddata/sdk-backend-spi";
 
 import { BearAuthenticatedCallGuard } from "../../../types/auth";
-import { convertWorkspaceUser } from "../../../convertors/toSdkModel/UsersConverter";
+import { convertWorkspaceUser } from "../../../convertors/fromBackend/UsersConverter";
 
 export class BearWorkspaceUsersQuery implements IWorkspaceUsersQuery {
     private options: IWorkspaceUsersQueryOptions = {};
