+++ libs/sdk-backend-bear/src/backend/workspaces/index.ts
@@ -1,6 +1,6 @@
 // (C) 2019-2020 GoodData Corporation
 import { IWorkspaceQueryFactory, IWorkspaceQuery, IWorkspaceQueryResult } from "@gooddata/sdk-backend-spi";
-import { convertUserProject } from "../../convertors/fromSdkModel/WorkspaceConverter";
+import { convertUserProject } from "../../convertors/toBackend/WorkspaceConverter";
 import { BearAuthenticatedCallGuard } from "../../types/auth";
 import { userLoginMd5FromAuthenticatedPrincipal } from "../../utils/api";
