+++ libs/sdk-backend-bear/src/convertors/fromBackend/CatalogConverter.ts
