+++ libs/sdk-backend-bear/src/convertors/fromBackend/DashboardConverter/widget.ts
