+++ libs/sdk-backend-bear/src/convertors/fromBackend/DateFilterConfigConverter.ts
