+++ libs/sdk-backend-bear/src/convertors/fromBackend/PropertiesConverter.ts
