+++ libs/sdk-backend-bear/src/convertors/fromBackend/VisualizationClassConverter.ts
