+++ libs/sdk-backend-bear/src/convertors/fromBackend/VisualizationConverter.ts
