+++ libs/sdk-backend-bear/src/convertors/toBackend/InsightConverter.ts
@@ -21,8 +21,8 @@ import {
 } from "@gooddata/sdk-model";
 import isEmpty from "lodash/isEmpty";
 import omitBy from "lodash/omitBy";
-import { convertUrisToReferences } from "../toSdkModel/ReferenceConverter";
-import { serializeProperties } from "../toSdkModel/PropertiesConverter";
+import { convertUrisToReferences } from "../fromBackend/ReferenceConverter";
+import { serializeProperties } from "../fromBackend/PropertiesConverter";
 import { convertExtendedFilter, shouldFilterBeIncluded } from "./FilterConverter";
 import { convertMeasure } from "./MeasureConverter";
