+++ libs/sdk-backend-bear/src/convertors/toBackend/MeasureConverter.ts
@@ -28,7 +28,7 @@ import {
     IPreviousPeriodDateDataSet,
 } from "@gooddata/sdk-model";
 import isEmpty from "lodash/isEmpty";
-import { toBearRef } from "../fromObjRef/ObjRefConverter";
+import { toBearRef } from "./ObjRefConverter";
 import { convertFilter, shouldFilterBeIncluded } from "./FilterConverter";
 
 const convertPreviousPeriodDataSet = (
