+++ libs/sdk-backend-bear/src/convertors/toBackend/afm/tests/FilterConverter.test.ts
