+++ libs/sdk-backend-bear/src/convertors/toBackend/afm/tests/InvalidInputs.fixture.ts
