+++ libs/sdk-backend-bear/src/utils/errorHandling.ts
